requirements.txt
streamlit==1.46.1
python-dotenv==1.1.1
pandas==2.3.1
pillow==11.3.0
matplotlib==3.10.3
matplotlib-venn==1.1.2
google-genai==1.25.0